requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.1
edgemanage==2.1.2
idna==2.10
importlib-metadata==1.7.0
ipaddr==2.2.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.0a1
Markdown==3.2.2
MarkupSafe==2.0.0a1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
setproctitle==1.1.10
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
